requirements.txt
Flask==2.3.3
# 移除mysqlclient，改用纯Python实现的PyMySQL
# mysqlclient==2.2.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
Jinja2==3.1.2
marshmallow==3.20.1
PyMySQL==1.1.0
flask-cors==4.0.0
gunicorn==21.2.0
cryptography>=42.0.0 